Add Windows ConPTY support for build, debug, and password detection
- build.zig: split Windows/Unix linking paths — Windows links pre-built
  libs from vendor/ghostty/zig-out/ to work around Zig 0.15 dependency
  system issues; add link_libcpp for C++ deps; add utfcpp.lib
- build.cmd: add clean/debug modes, loop-based lib copy, step numbering
- emacs.zig: fix vec API types from c_long to isize to match
  emacs-module.h ptrdiff_t (c_long is 32-bit on Windows LLP64)
- pty.zig: add comptime Windows guard for isPasswordMode (no POSIX API)
- ghostel.el: enable regex password detection on Windows where ConPTY
  has no termios fd
- ghostel-debug.el: add ConPTY spawn capture and platform-aware phase
  timing labels

+++ src/emacs.zig
@@ -160,15 +160,15 @@ pub const Env = struct {
 
     // --- Vectors ---
 
-    pub fn vecGet(self: Env, vec: Value, i: c_long) Value {
+    pub fn vecGet(self: Env, vec: Value, i: isize) Value {
         return self.raw.vec_get.?(self.raw, vec, i);
     }
 
-    pub fn vecSet(self: Env, vec: Value, i: c_long, value: Value) void {
+    pub fn vecSet(self: Env, vec: Value, i: isize, value: Value) void {
         self.raw.vec_set.?(self.raw, vec, i, value);
     }
 
-    pub fn vecSize(self: Env, vec: Value) c_long {
+    pub fn vecSize(self: Env, vec: Value) isize {
         return self.raw.vec_size.?(self.raw, vec);
     }
